trash/2/requirements.txt
pafy==0.5.5
python-vlc==3.0.18122
youtube-dl==2021.12.17
